Move more viewmodels into com.jerboa.model package
