+++ app/src/main/java/com/jerboa/model/AccountSettingsViewModel.kt
@@ -1,4 +1,4 @@
-package com.jerboa.ui.components.settings.account
+package com.jerboa.model
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -14,7 +14,6 @@ import com.jerboa.datatypes.types.LoginResponse
 import com.jerboa.datatypes.types.SaveUserSettings
 import com.jerboa.db.Account
 import com.jerboa.db.AccountRepository
-import com.jerboa.model.SiteViewModel
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
