+++ app/src/main/java/com/jerboa/ui/components/common/Navigation.kt
@@ -13,7 +13,7 @@ import com.jerboa.datatypes.types.Community
 import com.jerboa.datatypes.types.PostView
 import com.jerboa.datatypes.types.PrivateMessageView
 import com.jerboa.gson
-import com.jerboa.ui.components.comment.reply.ReplyItem
+import com.jerboa.model.ReplyItem
 
 // A view model stored higher up the tree used for moving navigation arguments from one route
 // to another. Since this will be reused, the value inside this should be moved out ASAP.
