+++ app/src/main/java/com/jerboa/ui/components/settings/account/AccountSettings.kt
@@ -29,6 +29,7 @@ import com.jerboa.datatypes.types.SaveUserSettings
 import com.jerboa.datatypes.types.SortType
 import com.jerboa.db.Account
 import com.jerboa.imageInputStreamFromUri
+import com.jerboa.model.AccountSettingsViewModel
 import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.components.common.*
 import com.jerboa.ui.theme.*
