+++ app/src/main/java/com/jerboa/ui/components/settings/account/AccountSettingsActivity.kt
@@ -12,6 +12,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavController
 import com.jerboa.R
 import com.jerboa.db.AccountViewModel
+import com.jerboa.model.AccountSettingsViewModel
 import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.components.common.SimpleTopAppBar
 import com.jerboa.ui.components.common.getCurrentAccount
